move contract update validation code to stdlib
