+++ runtime/errors.go
@@ -342,21 +342,6 @@ func (e *InvalidContractDeploymentError) Unwrap() error {
 	return e.Err
 }
 
-// ContractRemovalError
-//
-type ContractRemovalError struct {
-	Name string
-	interpreter.LocationRange
-}
-
-var _ errors.UserError = &ContractRemovalError{}
-
-func (*ContractRemovalError) IsUserError() {}
-
-func (e *ContractRemovalError) Error() string {
-	return fmt.Sprintf("cannot remove contract `%s`", e.Name)
-}
-
 // InvalidContractDeploymentOriginError
 //
 type InvalidContractDeploymentOriginError struct {
@@ -369,201 +354,4 @@ func (*InvalidContractDeploymentOriginError) IsUserError() {}
 
 func (*InvalidContractDeploymentOriginError) Error() string {
 	return "cannot deploy invalid contract"
-}
-
-// Contract update related errors
-
-// ContractUpdateError is reported upon any invalid update to a contract or contract interface.
-// It contains all the errors reported during the update validation.
-type ContractUpdateError struct {
-	ContractName string
-	Errors       []error
-	Location     common.Location
-}
-
-var _ errors.UserError = &ContractUpdateError{}
-var _ errors.ParentError = &ContractUpdateError{}
-
-func (*ContractUpdateError) IsUserError() {}
-
-func (e *ContractUpdateError) Error() string {
-	return fmt.Sprintf("cannot update contract `%s`", e.ContractName)
-}
-
-func (e *ContractUpdateError) ChildErrors() []error {
-	return e.Errors
-}
-
-func (e *ContractUpdateError) ImportLocation() common.Location {
-	return e.Location
-}
-
-// FieldMismatchError is reported during a contract update, when a type of a field
-// does not match the existing type of the same field.
-type FieldMismatchError struct {
-	DeclName  string
-	FieldName string
-	Err       error
-	ast.Range
-}
-
-var _ errors.UserError = &FieldMismatchError{}
-var _ errors.SecondaryError = &FieldMismatchError{}
-
-func (*FieldMismatchError) IsUserError() {}
-
-func (e *FieldMismatchError) Error() string {
-	return fmt.Sprintf("mismatching field `%s` in `%s`",
-		e.FieldName,
-		e.DeclName,
-	)
-}
-
-func (e *FieldMismatchError) SecondaryError() string {
-	return e.Err.Error()
-}
-
-// TypeMismatchError is reported during a contract update, when a type of the new program
-// does not match the existing type.
-type TypeMismatchError struct {
-	ExpectedType ast.Type
-	FoundType    ast.Type
-	ast.Range
-}
-
-var _ errors.UserError = &TypeMismatchError{}
-
-func (*TypeMismatchError) IsUserError() {}
-
-func (e *TypeMismatchError) Error() string {
-	return fmt.Sprintf("incompatible type annotations. expected `%s`, found `%s`",
-		e.ExpectedType,
-		e.FoundType,
-	)
-}
-
-// ExtraneousFieldError is reported during a contract update, when an updated composite
-// declaration has more fields than the existing declaration.
-type ExtraneousFieldError struct {
-	DeclName  string
-	FieldName string
-	ast.Range
-}
-
-var _ errors.UserError = &ExtraneousFieldError{}
-
-func (*ExtraneousFieldError) IsUserError() {}
-
-func (e *ExtraneousFieldError) Error() string {
-	return fmt.Sprintf("found new field `%s` in `%s`",
-		e.FieldName,
-		e.DeclName,
-	)
-}
-
-// ContractNotFoundError is reported during a contract update, if no contract can be
-// found in the program.
-type ContractNotFoundError struct {
-	ast.Range
-}
-
-var _ errors.UserError = &ContractNotFoundError{}
-
-func (*ContractNotFoundError) IsUserError() {}
-
-func (e *ContractNotFoundError) Error() string {
-	return "cannot find any contract or contract interface"
-}
-
-// InvalidDeclarationKindChangeError is reported during a contract update, when an attempt is made
-// to convert an existing contract to a contract interface, or vise versa.
-type InvalidDeclarationKindChangeError struct {
-	Name    string
-	OldKind common.DeclarationKind
-	NewKind common.DeclarationKind
-	ast.Range
-}
-
-var _ errors.UserError = &InvalidDeclarationKindChangeError{}
-
-func (*InvalidDeclarationKindChangeError) IsUserError() {}
-
-func (e *InvalidDeclarationKindChangeError) Error() string {
-	return fmt.Sprintf("trying to convert %s `%s` to a %s", e.OldKind.Name(), e.Name, e.NewKind.Name())
-}
-
-// ConformanceMismatchError is reported during a contract update, when the enum conformance of the new program
-// does not match the existing one.
-type ConformanceMismatchError struct {
-	DeclName string
-	ast.Range
-}
-
-var _ errors.UserError = &ConformanceMismatchError{}
-
-func (*ConformanceMismatchError) IsUserError() {}
-
-func (e *ConformanceMismatchError) Error() string {
-	return fmt.Sprintf("conformances does not match in `%s`", e.DeclName)
-}
-
-// EnumCaseMismatchError is reported during an enum update, when an updated enum case
-// does not match the existing enum case.
-type EnumCaseMismatchError struct {
-	ExpectedName string
-	FoundName    string
-	ast.Range
-}
-
-var _ errors.UserError = &EnumCaseMismatchError{}
-
-func (*EnumCaseMismatchError) IsUserError() {}
-
-func (e *EnumCaseMismatchError) Error() string {
-	return fmt.Sprintf("mismatching enum case: expected `%s`, found `%s`",
-		e.ExpectedName,
-		e.FoundName,
-	)
-}
-
-// MissingEnumCasesError is reported during an enum update, if any enum cases are removed
-// from an existing enum.
-type MissingEnumCasesError struct {
-	DeclName string
-	Expected int
-	Found    int
-	ast.Range
-}
-
-var _ errors.UserError = &MissingEnumCasesError{}
-
-func (*MissingEnumCasesError) IsUserError() {}
-
-func (e *MissingEnumCasesError) Error() string {
-	return fmt.Sprintf(
-		"missing cases in enum `%s`: expected %d or more, found %d",
-		e.DeclName,
-		e.Expected,
-		e.Found,
-	)
-}
-
-// MissingDeclarationError is reported during a contract update,
-// if an existing declaration is removed.
-type MissingDeclarationError struct {
-	Name string
-	Kind common.DeclarationKind
-	ast.Range
-}
-
-var _ errors.UserError = &MissingDeclarationError{}
-
-func (*MissingDeclarationError) IsUserError() {}
-
-func (e *MissingDeclarationError) Error() string {
-	return fmt.Sprintf(
-		"missing %s declaration `%s`",
-		e.Kind,
-		e.Name,
-	)
 }
